Adding infos about installation and Usage
Hi !
I runned you code locally because I want to use parts of it for a project.

This PR is just to add some informations that could be useful for people wanting to run your code. 
Feel free to edit it. Or not merge if you don't want it.
Hope it helps,
Mathis